[!] Version 1..
Okay, that's the first essential version and I'm happy with it.

What [me_Console] does

It allows you to print values of basic types without thinking of their
representation. They will be represented in some consistent way
and you'd better love the way they are represented. Or modify
implementation to suit your taste.

Important part for me is indentation. I want ability to print indented
values. Now it's as simple as

  Console.Indent();
  Console.Print("Here we are.");
  Console.Unindent();

Also fine qol thing is

  Console.Line()

which just prints "--" on new line.

"--" looks too short. But just making this constant string longer
is bad decision. We need <Console.Width> conception.

And that console width should be respected by data printers. They
should wrap to next indented newline when needed.

And that is what I'm going to implement in next generation.