[-] Removed Flush()..
That was not needed.

Problem with first implementation was that

  Print("Hello")

printed

  Hello_

leaving cursor on line until next Print() or Write(). Which is really
annoying when you printing your output and waiting for user's command.

Later I decided that Print() for strings is always prints tail newline.

Current code possibly can be simplified with this contract change.